backend: add game challenges
Whenever a user wants to create a game, they'll start by creating a game
challenge. The challenge encapsulates the various game parameters, and
can either be public (i.e. should be listed on the main site for others
to accept) or private (i.e. generates an invitation URL that the user
can send to a friend).

Challenges also default to expiring in 1 day, and we should eventually
add some sort of regular check for expired challenges..